More test/impl refactoring
