+++ src/main/java/tools/jackson/databind/node/BaseJsonNode.java
@@ -140,7 +140,25 @@ public abstract class BaseJsonNode
         // Overridden by NumericNode, for other types return default
         return Optional.empty();
     }
-    
+
+    /*
+    @Override
+    public BigInteger asBigInteger() {
+        return _reportCoercionFail("asBigInteger()", BigInteger.class, "value type not numeric");
+    }
+
+    @Override
+    public BigInteger asBigInteger(BigInteger defaultValue) {
+        // Overridden by NumericNode, for other types return default
+        return defaultValue;
+    }
+
+    @Override
+    public Optional<BigInteger> asBigIntegerOpt() {
+        // Overridden by NumericNode, for other types return default
+        return Optional.empty();
+    }
+    */
     @Override
     public float floatValue() {
         return _reportCoercionFail("floatValue()", Float.TYPE, "value type not numeric");
